refactor(engine): decompose local_copy/executor/file/copy/transfer/execute.rs
Split the 1013-line execute.rs into a directory of focused submodules:
- mod.rs (orchestration entry)
- skip.rs (skip + metadata-reuse paths)
- clonefile.rs (macOS clonefile fast path + normalize)
- iouring.rs (Linux io_uring dispatch)

Public API preserved via re-exports from mod.rs.
